providedPosition in hooker now has right value

+++ src/compressedsection.hpp
@@ -3,8 +3,8 @@
 
 struct CompressedSection
 {
-    uint32_t Start;
-    uint32_t End;
+    size_t Start;
+	size_t End;
     uint8_t Size;
 
     CompressedSection(uint32_t start, uint32_t end, uint8_t size)
